refactor almost done
two more things 1. he seems to be jumping up the slope when jumping and hiting the slope at a specific angle            2. when he is sliding down a slope and ray hits the ground it kills momentumn before he hits the ground thus stoping him mid air and then he continiues falling (might be cuz of infinity range on a ray in descendingSlope)